Updated Algorithm, Fixed Bug
Updated the algorithm for the signature.

Fixed a bug where the download button wouldn't appear for videos that
were being played in a playlist or when you clicked on related videos.
This is because youtube wasn't doing a full page refresh. Now
setinterval just runs indefinitely checking every 500ms to see if the
YTD button is there.